订单服务 1.2

+++ oreder/src/main/java/com/example/order/model/bo/NewOrder.java
@@ -28,7 +28,7 @@ public class NewOrder implements VoObject {
     private String address;
     private String mobile;
     private String message;
-    private int orderType;
+    private Byte orderType;
     private Long freightPrice;
     private Long couponId;
     private Long couponActivityId;
@@ -39,9 +39,9 @@ public class NewOrder implements VoObject {
     private Integer rebateNum;
     private LocalDateTime confirmTime;
     private String shipmentSn;
-    private int state;
-    private int substate;
-    private int beDeleted;
+    private Byte state;
+    private Byte substate;
+    private Byte beDeleted;
     private LocalDateTime gmtCreate;
     private LocalDateTime gmtModified;
     private Long grouponId;
